multiple minishell solved
