+++ output.txt
@@ -1,1 +1,1 @@
-hello world axel 
+hello world
